Move ApiBacklogItems schema
