+++ backlog/collection/index.js
@@ -1,76 +1,6 @@
-import { ApiBackends } from '/apis/collection/backend';
-
 const ApiBacklogItems = new Mongo.Collection("apiBacklogItems");
 
 export { ApiBacklogItems };
-
-ApiBacklogItems.attachSchema(new SimpleSchema({
-  title: {
-    type: String,
-    label: "Title",
-    max: 100,
-    autoform: {
-      placeholder: "Title"
-    }
-  },
-  details: {
-    type: String,
-    label: "Details",
-    max: 1000,
-    autoform: {
-      rows: 5,
-      placeholder: "Description"
-    }
-  },
-  priority: {
-    type: Number,
-    label: 'Priority',
-    min:0,
-    max:2,
-    autoform: {
-      options: [
-        { label: "High", value: 2 },
-        { label: "Middle", value: 1 },
-        { label: "None", value: 0 }
-      ]
-    }
-  },
-  userId: {
-    type: String,
-    regEx: SimpleSchema.RegEx.Id,
-    autoValue: function () {
-      return Meteor.userId();
-    }
-  },
-  apiBackendId: {
-    type: String,
-    regEx: SimpleSchema.RegEx.Id,
-    optional: true
-  },
-  createdAt: {
-    type: Date,
-    autoValue: function () {
-
-      // Check if mongoDB insert operation is initial
-      if (this.isInsert)
-        return new Date();
-
-      // Check if mongoDB insert operation is initial
-      else if (this.isUpsert)
-        return { $setOnInsert: new Date() };
-
-      // If not - field is not updated
-      else
-        this.unset();
-    }
-  },
-  updatedAt: {
-    type: Date,
-    autoValue: function () {
-      return new Date();
-    }
-  }
-}));
 
 ApiBacklogItems.allow({
   insert: function (userId, backlog) {
